tools/requirements.txt
pyswisseph>=2.10.0
tabulate>=0.9.0
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
# NotebookLM連携（ローカル実行用）
notebooklm-py>=0.3.4
